+++ R/cooccurrences.R
@@ -15,12 +15,12 @@ NULL
 #' @rdname cooccurrences-class
 setMethod('summary', 'cooccurrences', function(object) {
   cat("\n** Context object: **\n")
-  cat(sprintf("%-20s", "CWB-Korpus:"), object@corpus, "\n")
-  cat(sprintf("%-20s", "Partition:"), object@partition, "\n")
-  cat(sprintf("%-20s", "Node:"), object@query, "\n")
-  cat(sprintf("%-20s", "P-Attribute:"), object@p_attribute, "\n")
-  cat(sprintf("%-20s", "Node count:"), object@count, "\n")
-  cat(sprintf("%-20s", "Stat table length:"), nrow(object@stat), "\n\n")
+  cat(sprintf("%-20s", "CWB corpus:"), object@corpus, "\n")
+  cat(sprintf("%-20s", "partition:"), object@partition, "\n")
+  cat(sprintf("%-20s", "node:"), object@query, "\n")
+  cat(sprintf("%-20s", "p-Attribute:"), object@p_attribute, "\n")
+  cat(sprintf("%-20s", "node count:"), object@count, "\n")
+  cat(sprintf("%-20s", "stat table length:"), nrow(object@stat), "\n\n")
 })