+++ R/use.R
@@ -50,10 +50,11 @@ use <- function(pkg = NULL, lib.loc = .libPaths(), dir = getOption("polmineR.def
         registry_new = registry(),
         home_dir_new = system.file("extdata", "cwb", "indexed_corpora", tolower(corpus), package = pkg, lib.loc = lib.loc)
         )
+      setTemplate(toupper(corpus))
     }
-    registry_reset(registryDir = registry())
-    options("polmineR.templates" = list())
-    setTemplate()
+    # registry_reset(registryDir = registry())
+    # options("polmineR.templates" = list())
+    # setTemplate()
   } else {
     if (dir.exists(dir)){
       previousRegistry <- registry_reset(dir, verbose = verbose)