+++ NEWS.md
@@ -1,5 +1,5 @@
 # v0.7.9
-* new as.list,bundle-method for convenience
+* new as.list,bundle-method for convenience, to access slot objects
 * as.bundle is more generic now, so that any kind of object can be coerced to a bundle now
 * as.speeches-method turned into function that allows partition and corpus as input
 * is.partition-function introduced
@@ -14,7 +14,10 @@
 * new partition_bundle,partition_bundle-method introduced
 * naming of methods and functions, classes and most arguments moved to snake_case, maintaining backwards compatibility
 * utility function getObjects not exported any more
-
+* for count,partition_bundle-method, column 'partition' will be a character vector now (not factor)
+* new argument 'type' added to partition_bundle
+* new method 'get_type' introduced to make getting corpus type more robust
+* bug removed that has caused a crash when cutoff is larger than number of tokens in a partition when calling get_token_stream
 
 # v0.7.8
 * upon loading the package, check that data directories are set correctly in registry files to make sure that sample data in pre-compiled packages can be used